Separate characterDefaults into its own file
