+++ src/core/shared/interfaces/character.ts
@@ -1,5 +1,4 @@
 import { Vector3 } from './vector';
-import { SHARED_CONFIG } from '../configurations/shared';
 import { Appearance } from './appearance';
 import { CharacterInfo } from './characterInfo';
 import { Item } from './item';
@@ -169,18 +168,3 @@ export interface Character {
      */
     faction: string | null;
 }
-
-export const CharacterDefaults: Partial<Character> = {
-    pos: SHARED_CONFIG.PLAYER_NEW_SPAWN_POS as Vector3,
-    cash: SHARED_CONFIG.PLAYER_CASH,
-    bank: SHARED_CONFIG.PLAYER_BANK,
-    appearance: {},
-    info: {},
-    food: 100,
-    water: 100,
-    isDead: false,
-    health: 199,
-    armour: 0,
-    hours: 0,
-    wanted: 0,
-};
