+++ src/core/shared/interfaces/index.ts
@@ -2,7 +2,8 @@ export * from './actions';
 export * from './animation';
 export * from './appearance';
 export * from './blip';
-export { Character, CharacterDefaults } from './character';
+export * from './character';
+export * from './characterDefaults';
 export * from './characterInfo';
 export * from './clothing';
 export * from './command';
